+++ R/AFM.histogram.R
@@ -6,12 +6,16 @@
 #' @param dataOnly if \code{TRUE} a data frame with the histogram
 #' data is returned
 #' @return data frame or ggplot
+#'
 #' @author Thomas Gredig
+#'
 #' @examples
 #' d = AFM.import(AFM.getSampleImages(type='ibw')[1])
 #' head(AFM.histogram(d, dataOnly=TRUE),n=20)
 #' AFM.histogram(d)
 #' @export
+#' @seealso \code{\link{plot.AFMdata}}
+#'
 #' @importFrom ggplot2 ggplot aes geom_histogram geom_density
 #' @importFrom graphics hist
 AFM.histogram <- function(obj, no=1, binNo = 200, dataOnly=FALSE) {